Test (#1)
* spiffs Test

* feat: Add wifiManager class for managing WiFi connection

This commit adds a new class called wifiManager to handle WiFi connection management. The class includes methods for inputting the SSID and password, connecting to the WiFi network, and getting the connection status. This will be useful for configuring and managing WiFi settings in the application.

* feat: Add misc menu to App class

This commit adds a new menu option to the App class called "misc". The misc menu allows the user to perform various miscellaneous tasks, such as saving and loading profiles. This feature enhances the functionality of the application by providing additional options for managing settings and configurations.

* send data to beebotte